pyproject.toml
[tool.poetry]
name = "concurrencymq"
version = "0.1.0"
description = ""
authors = ["Edwin Millan <bassgs3000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyzmq = "^25.1.2"
tornado = "^6.4"
pynacl = "^1.5.0"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
httpx = {extras = ["http2"], version = "^0.26.0"}
hypercorn = "^0.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
